+++ Changelog.md
@@ -1,9 +1,17 @@
 ## Version 0.5 Changelog
 * **Added**
-  * Jade Support
-  * Tree will only fall if a tool is used (Configurable)
+  * [Jade](https://www.curseforge.com/minecraft/mc-mods/jade) Support
+  * Tree will only fall if a tool is used (Configurable | Default false)
   * Tree Size Limiter (Configurable)
+  * Single block mining method (Client | Configurable | Possible: Crouch, Keybind toggle, off [off means you will have to set if your mining a single block through the config] | Default Crouch)
+  * Block mining speed is slower for bigger trees (Configurable)
+  * Tree fall sound effect
+
 
 * **Changes**
   * Change food exhaustion to be configurable
   * Change item damage to be configurable
+
+
+* **known issues**
+  * Tree mining speed won't change back to the default mining speed when single block mining is active.